several enhancement
